# Conflicts:
#	src/main/java/pl/kuba/domain/servises/CarService.java
#	src/main/java/pl/kuba/infrastructure/persistence/BranchRepository.java
#	src/main/java/pl/kuba/infrastructure/persistence/ReservationRepository.java

+++ src/main/java/pl/kuba/infrastructure/persistence/RentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.kuba.entities.Rent;
 
 @Repository
-public interface RentRepository extends JpaRepository<Rent,Integer> {
+public interface RentRepository extends JpaRepository<Rent,Long> {
 }
