+++ src/main/java/pl/kuba/infrastructure/persistence/RevenueRepository.java
@@ -3,5 +3,5 @@ package pl.kuba.infrastructure.persistence;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.kuba.entities.Revenue;
 
-public interface RevenueRepository extends JpaRepository<Revenue,Integer> {
+public interface RevenueRepository extends JpaRepository<Revenue,Long> {
 }
